[REFACTOR] 날짜, 달력 유틸 함수 분리 후 적용
