+++ src/app/mypage/page.tsx
@@ -5,7 +5,7 @@ import Button from "@/components/designSystem/Button";
 import MyInfo from "@/components/mypage/article/myInfo/MyInfo";
 import MySettings from "@/components/mypage/article/MySettings";
 import MyWritten from "@/components/mypage/article/MyWritten";
-import Mydiscussions from "@/components/mypage/article/Mydiscussions";
+import Mydiscussions from "@/components/mypage/article/MyDiscussion/Mydiscussions";
 import { BG_COLOR, TEXT_COLOR } from "@/constants/global/colors";
 import { MyNeighbors } from "@/components/mypage/article/neighbors/MyNeighbors";
 import withAuth from "@/components/provider/withAuth";
