+++ src/components/mypage/TimeOfLearning.tsx
@@ -1,13 +1,10 @@
 import React from "react";
 
-import getTime from "@/utils/util-getTime";
-
 import Button from "../designSystem/Button";
 
 const TimeOfLearning: React.FC<{ category: timeOfLearning }> = ({
   category,
 }) => {
-  const timeOfLearning = getTime(category.timespent);
   return (
     <div
       key={`mypage${category.category}`}
@@ -22,7 +19,7 @@ const TimeOfLearning: React.FC<{ category: timeOfLearning }> = ({
       >
         # {category.category}
       </Button>
-      <span>{timeOfLearning}</span>
+      <span>{category.timespent}</span>
     </div>
   );
 };
